requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
Jinja2==3.1.4
apscheduler==3.10.4
